Fix navigation of create account flow
