+++ src/hooks.ts
@@ -5,7 +5,7 @@ import type { RootState, AppDispatch } from './store';
 export const useAppDispatch = () => useDispatch<AppDispatch>();
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 
-export const useIsAuthenticated = () => {
+export const useUserSelector = () => {
   return useAppSelector(
     state =>
       !!(state.account.signupAccount?.uid || state.account.loginAccount?.uid),
